Use better resolution for pts computation. Fixes desync in Real live streams with huge pts values
git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@19701 b3059339-0415-0410-9bf9-f77b7e298cf2

+++ libmpdemux/video.c
@@ -607,7 +607,7 @@ int video_read_frame(sh_video_t* sh_video,float* frame_time_ptr,unsigned char** 
       break;
       case DEMUXER_TYPE_REAL:
         {
-          float next_pts = ds_get_next_pts(d_video);
+          double next_pts = ds_get_next_pts(d_video);
           float d = next_pts > 0 ? next_pts - d_video->pts : d_video->pts - pts1;
 
           frame_time = (d >= 0 && pts1 > 0) ? d : 0.001;
